[feat] 메뉴 레포 업데이트
